refactor: Timeline.tsx 분리 (M3)
- timeline/constants.ts: 상수 + 타입 (24줄)
- timeline/useDrag.ts: 드래그 로직 커스텀 훅 (302줄)
- timeline/useKeyboardShortcuts.ts: 키보드 단축키 훅 (85줄)
- timeline/useWaveform.ts: 파형 추출/렌더링 훅 (174줄)
- Timeline.tsx: 1,311줄 → 787줄 (메인 컴포넌트)
