+++ dashboard/src/components/studio/Timeline.tsx
@@ -1,27 +1,15 @@
 'use client';
 
-// TODO: 트랙 렌더링(TrackRow), 드래그 로직(useDrag hook), 키보드 단축키(useKeyboardShortcuts hook) 분리 예정
-// TODO: useShallow로 관련 selector 그룹핑하여 리렌더링 최적화
-
 import React, { useRef, useCallback, useState, useEffect } from 'react';
 import { getEditorConfig } from '@/lib/editor-config';
 import { useEditorStore } from './store';
 import type { Clip } from './types';
 import { NUM_VIDEO_TRACKS } from './types';
 
-const FPS = 30;
-const TRACK_HEIGHT = 48;
-const CLIP_AUDIO_TRACK_HEIGHT = 32;
-const HEADER_W = 60;
-const TRIM_HANDLE_W = 6;
-const MIN_CLIP_DURATION = 0.1;
-const SNAP_THRESHOLD_PX = 8;
-
-type ClipAudioInfo = {
-  hasAudio: boolean;
-};
-
-type DragMode = null | 'playhead' | 'clip-move' | 'clip-trim-left' | 'clip-trim-right' | 'sub-move' | 'sub-trim-left' | 'sub-trim-right' | 'bgm-move' | 'bgm-trim-left' | 'bgm-trim-right';
+import { FPS, TRACK_HEIGHT, CLIP_AUDIO_TRACK_HEIGHT, HEADER_W, TRIM_HANDLE_W } from './timeline/constants';
+import { useWaveform } from './timeline/useWaveform';
+import { useKeyboardShortcuts } from './timeline/useKeyboardShortcuts';
+import { useDrag } from './timeline/useDrag';
 
 export const Timeline: React.FC = () => {
   const clips = useEditorStore((s) => s.clips);
@@ -39,7 +27,6 @@ export const Timeline: React.FC = () => {
   const setSelectedClipIndex = useEditorStore((s) => s.setSelectedClipIndex);
   const toggleClipSelection = useEditorStore((s) => s.toggleClipSelection);
   const selectClipRange = useEditorStore((s) => s.selectClipRange);
-  const moveClips = useEditorStore((s) => s.moveClips);
   const setSelectedSubIndex = useEditorStore((s) => s.setSelectedSubIndex);
   const setActivePanel = useEditorStore((s) => s.setActivePanel);
   const setPxPerSec = useEditorStore((s) => s.setPxPerSec);
@@ -47,77 +34,20 @@ export const Timeline: React.FC = () => {
   const bgmEnabled = useEditorStore((s) => s.bgmEnabled);
   const removeClip = useEditorStore((s) => s.removeClip);
   const removeSubtitle = useEditorStore((s) => s.removeSubtitle);
-  const updateClip = useEditorStore((s) => s.updateClip);
   const updateClipMeta = useEditorStore((s) => s.updateClipMeta);
   const setAllClipAudioMuted = useEditorStore((s) => s.setAllClipAudioMuted);
-  const updateGlobalSub = useEditorStore((s) => s.updateGlobalSub);
-  const updateBgmClip = useEditorStore((s) => s.updateBgmClip);
-  const moveClip = useEditorStore((s) => s.moveClip);
   const splitClip = useEditorStore((s) => s.splitClip);
   const undo = useEditorStore((s) => s.undo);
   const redo = useEditorStore((s) => s.redo);
   const canUndo = useEditorStore((s) => s.canUndo);
   const canRedo = useEditorStore((s) => s.canRedo);
-  const copySubtitle = useEditorStore((s) => s.copySubtitle);
-  const copyClip = useEditorStore((s) => s.copyClip);
-  const paste = useEditorStore((s) => s.paste);
   const addClip = useEditorStore((s) => s.addClip);
-  const moveClipToTrackAndTime = useEditorStore((s) => s.moveClipToTrackAndTime);
 
   const [fileDragOver, setFileDragOver] = useState(false);
-  const handleFileDrop = useCallback(async (files: FileList) => {
-    const videoExts = ['.mp4', '.mov', '.mkv', '.avi', '.webm'];
-    for (const file of Array.from(files)) {
-      const ext = file.name.slice(file.name.lastIndexOf('.')).toLowerCase();
-      if (!videoExts.includes(ext)) continue;
-      try {
-        const formData = new FormData();
-        formData.append('file', file);
-        // Next.js rewrite 프록시 먼저 시도, 실패 시 직접 요청
-        let uploadOk = false;
-        try {
-          const res = await fetch(`${getEditorConfig().apiUrl}/api/upload`, { method: 'POST', body: formData });
-          uploadOk = res.ok;
-        } catch {}
-        if (!uploadOk) {
-          const formData2 = new FormData();
-          formData2.append('file', file);
-          const res2 = await fetch(`${getEditorConfig().apiUrl}/api/upload`, { method: 'POST', body: formData2 });
-          uploadOk = res2.ok;
-        }
-        if (!uploadOk) {
-          console.error('[드래그앤드롭] 업로드 실패');
-          continue;
-        }
-        let duration = 10;
-        for (let retry = 0; retry < 3; retry++) {
-          try {
-            const pr = await fetch(`${getEditorConfig().apiUrl}/api/media/probe/${encodeURIComponent(file.name)}`);
-            if (pr.ok) {
-              const pd = await pr.json();
-              if (pd.duration && pd.duration > 0) { duration = pd.duration; break; }
-            }
-          } catch {}
-          await new Promise((r) => setTimeout(r, 500));
-        }
-        addClip({ source: file.name, start: 0, end: duration, source_idx: 0 } as Clip);
-      } catch (e) {
-        console.error('[타임라인 드래그앤드롭] 영상 추가 실패:', e);
-      }
-    }
-  }, [addClip]);
+  const [selectedBgmIndex, setSelectedBgmIndex] = useState(-1);
 
   const trackRef = useRef<HTMLDivElement>(null);
   const scrollContainerRef = useRef<HTMLDivElement>(null);
-  const [dragMode, setDragMode] = useState<DragMode>(null);
-  const [dragIndex, setDragIndex] = useState(-1);
-  const dragStartX = useRef(0);
-  const dragStartVal = useRef<{ start: number; end: number; audioStart?: number }>({ start: 0, end: 0 });
-  const dragOffsetSec = useRef(0); // clip-move: 드래그 시작 시 마우스와 클립 왼쪽 간 오프셋(초)
-
-  // Snap indicator line
-  const [snapLineX, setSnapLineX] = useState<number | null>(null);
-  const [selectedBgmIndex, setSelectedBgmIndex] = useState(-1);
 
   // Calculate clip positions on timeline
   const clipDurations = clips.map((clip, i) => {
@@ -137,334 +67,60 @@ export const Timeline: React.FC = () => {
   const currentTime = currentFrame / FPS;
   const allClipAudioMuted = clips.length > 0 && clips.every((_, i) => clipMeta[i]?.audioMuted ?? false);
 
-  // Snap points: ALL element boundaries + playhead (Feature 3)
-  const snapPoints: number[] = [];
-  clips.forEach((c, i) => {
-    const cs = c.timelineStart ?? clipStarts[i];
-    snapPoints.push(cs);
-    snapPoints.push(cs + clipDurations[i]);
-  });
-  globalSubs.forEach((sub) => {
-    snapPoints.push(sub.start);
-    snapPoints.push(sub.end);
-  });
-  bgmClips.forEach((b) => {
-    snapPoints.push(b.start);
-    snapPoints.push(b.start + b.duration);
-  });
-  // Also snap to playhead
-  snapPoints.push(currentTime);
+  // Waveform hook
+  const { bgmWaveformDurations, clipAudioInfo, waveformVersion, drawBgmWaveform, drawClipWaveform } = useWaveform();
 
-  const snapToNearest = (time: number, excludePoints?: number[]): { time: number; snapped: boolean } => {
-    const threshold = SNAP_THRESHOLD_PX / pxPerSec;
-    let best = time;
-    let bestDist = threshold;
-    let snapped = false;
-    for (const sp of snapPoints) {
-      if (excludePoints && excludePoints.some((ep) => Math.abs(ep - sp) < 0.001)) continue;
-      const dist = Math.abs(time - sp);
-      if (dist < bestDist) {
-        bestDist = dist;
-        best = sp;
-        snapped = true;
+  // Drag hook
+  const { dragMode, dragIndex, dragOffsetSec, snapLineX, startDrag, handleTimelineClick, getTimeFromX } = useDrag({
+    clipStarts,
+    clipDurations,
+    trackRef,
+    bgmWaveformDurations,
+  });
+
+  // Keyboard shortcuts hook
+  useKeyboardShortcuts(selectedBgmIndex, setSelectedBgmIndex);
+
+  // File drop handler
+  const handleFileDrop = useCallback(async (files: FileList) => {
+    const videoExts = ['.mp4', '.mov', '.mkv', '.avi', '.webm'];
+    for (const file of Array.from(files)) {
+      const ext = file.name.slice(file.name.lastIndexOf('.')).toLowerCase();
+      if (!videoExts.includes(ext)) continue;
+      try {
+        const formData = new FormData();
+        formData.append('file', file);
+        let uploadOk = false;
+        try {
+          const res = await fetch(`${getEditorConfig().apiUrl}/api/upload`, { method: 'POST', body: formData });
+          uploadOk = res.ok;
+        } catch { /* ignore */ }
+        if (!uploadOk) {
+          const formData2 = new FormData();
+          formData2.append('file', file);
+          const res2 = await fetch(`${getEditorConfig().apiUrl}/api/upload`, { method: 'POST', body: formData2 });
+          uploadOk = res2.ok;
+        }
+        if (!uploadOk) continue;
+        let duration = 10;
+        for (let retry = 0; retry < 3; retry++) {
+          try {
+            const pr = await fetch(`${getEditorConfig().apiUrl}/api/media/probe/${encodeURIComponent(file.name)}`);
+            if (pr.ok) {
+              const pd = await pr.json();
+              if (pd.duration && pd.duration > 0) { duration = pd.duration; break; }
+            }
+          } catch { /* ignore */ }
+          await new Promise((r) => setTimeout(r, 500));
+        }
+        addClip({ source: file.name, start: 0, end: duration, source_idx: 0 } as Clip);
+      } catch {
+        // ignore
       }
     }
-    return { time: best, snapped };
-  };
+  }, [addClip]);
 
-  // Seek
-  const seekToTime = useCallback((time: number) => {
-    const frame = Math.max(0, Math.round(time * FPS));
-    const seekTo = (window as unknown as Record<string, (f: number) => void>).__studioSeekTo;
-    if (seekTo) seekTo(frame);
-  }, []);
-
-  const handleTimelineClick = useCallback((e: React.MouseEvent<HTMLDivElement>) => {
-    if (dragMode) return;
-    const rect = (trackRef.current || e.currentTarget).getBoundingClientRect();
-    const x = e.clientX - rect.left;
-    if (x < 0) return;
-    seekToTime(x / pxPerSec);
-  }, [pxPerSec, seekToTime, dragMode]);
-
-  // Drag handlers
-  const getTimeFromX = useCallback((clientX: number) => {
-    if (!trackRef.current) return 0;
-    const rect = trackRef.current.getBoundingClientRect();
-    return Math.max(0, (clientX - rect.left) / pxPerSec);
-  }, [pxPerSec]);
-
-  const handleMouseMove = useCallback((e: MouseEvent) => {
-    if (!dragMode) return;
-    const deltaX = e.clientX - dragStartX.current;
-    const deltaSec = deltaX / pxPerSec;
-
-    if (dragMode === 'playhead') {
-      seekToTime(getTimeFromX(e.clientX));
-      setSnapLineX(null);
-    } else if (dragMode === 'clip-trim-left') {
-      const speed = clipMeta[dragIndex]?.speed ?? 1;
-      const rawStart = Math.max(0, dragStartVal.current.start + deltaSec * speed);
-      if (rawStart < dragStartVal.current.end - MIN_CLIP_DURATION) {
-        // 타임라인 끝 위치 고정: timelineStart를 조절해서 끝 위치 유지
-        const origTimelineStart = dragStartVal.current.audioStart ?? (clips[dragIndex].timelineStart ?? clipStarts[dragIndex]);
-        const origTimelineEnd = origTimelineStart + (dragStartVal.current.end - dragStartVal.current.start) / speed;
-        const sourceDelta = rawStart - dragStartVal.current.start;
-        const newTimelineStart = origTimelineStart + sourceDelta / speed;
-        // Snap the timeline edge of the clip's left side
-        const snap = snapToNearest(newTimelineStart, [origTimelineStart]);
-        if (snap.snapped) {
-          const snappedSourceDelta = (snap.time - origTimelineStart) * speed;
-          const snappedStart = dragStartVal.current.start + snappedSourceDelta;
-          if (snappedStart >= 0 && snappedStart < dragStartVal.current.end - MIN_CLIP_DURATION) {
-            updateClip(dragIndex, { start: snappedStart, timelineStart: snap.time });
-            setSnapLineX(snap.time * pxPerSec);
-          }
-        } else {
-          updateClip(dragIndex, { start: rawStart, timelineStart: Math.max(0, newTimelineStart) });
-          setSnapLineX(null);
-        }
-      }
-    } else if (dragMode === 'clip-trim-right') {
-      const speed = clipMeta[dragIndex]?.speed ?? 1;
-      let newEnd = Math.max(dragStartVal.current.start + MIN_CLIP_DURATION, dragStartVal.current.end + deltaSec * speed);
-      // Snap the timeline edge of the clip's right side
-      const clipTlStart = clips[dragIndex].timelineStart ?? clipStarts[dragIndex];
-      const newTimelineEnd = clipTlStart + (newEnd - clips[dragIndex].start) / speed;
-      const origTimelineEnd = clipTlStart + clipDurations[dragIndex];
-      const snap = snapToNearest(newTimelineEnd, [origTimelineEnd]);
-      if (snap.snapped) {
-        newEnd = clips[dragIndex].start + (snap.time - clipTlStart) * speed;
-        newEnd = Math.max(dragStartVal.current.start + MIN_CLIP_DURATION, newEnd);
-        updateClip(dragIndex, { end: newEnd });
-        setSnapLineX(snap.time * pxPerSec);
-      } else {
-        updateClip(dragIndex, { end: newEnd });
-        setSnapLineX(null);
-      }
-    } else if (dragMode === 'clip-move') {
-      // 절대 위치 이동: 수평(timelineStart) + 수직(track)
-      const rawTime = getTimeFromX(e.clientX) - dragOffsetSec.current;
-      const newTimelineStart = Math.max(0, rawTime);
-
-      // 수직 → track 결정
-      if (!trackRef.current) return;
-      const trackRect = trackRef.current.getBoundingClientRect();
-      const relY = e.clientY - trackRect.top;
-      // 트랙 레이아웃: 트랙2(상단)→트랙1→트랙0(하단), 각 TRACK_HEIGHT 높이
-      let newTrack = Math.floor(relY / TRACK_HEIGHT);
-      newTrack = NUM_VIDEO_TRACKS - 1 - newTrack; // 반전: 상단=트랙2, 하단=트랙0
-      newTrack = Math.max(0, Math.min(NUM_VIDEO_TRACKS - 1, newTrack));
-
-      // 스냅
-      const excludePoints = [clips[dragIndex].timelineStart ?? 0, (clips[dragIndex].timelineStart ?? 0) + clipDurations[dragIndex]];
-      const snappedStart = snapToNearest(newTimelineStart, excludePoints);
-      const snappedEnd = snapToNearest(newTimelineStart + clipDurations[dragIndex], excludePoints);
-
-      let finalStart = newTimelineStart;
-      if (snappedStart.snapped && (!snappedEnd.snapped || Math.abs(snappedStart.time - newTimelineStart) <= Math.abs(snappedEnd.time - (newTimelineStart + clipDurations[dragIndex])))) {
-        finalStart = snappedStart.time;
-        setSnapLineX(snappedStart.time * pxPerSec);
-      } else if (snappedEnd.snapped) {
-        finalStart = snappedEnd.time - clipDurations[dragIndex];
-        setSnapLineX(snappedEnd.time * pxPerSec);
-      } else {
-        setSnapLineX(null);
-      }
-      finalStart = Math.max(0, finalStart);
-
-      moveClipToTrackAndTime(dragIndex, newTrack, finalStart);
-    } else if (dragMode === 'sub-move') {
-      const sub = globalSubs[dragIndex];
-      if (!sub) return;
-      const dur = dragStartVal.current.end - dragStartVal.current.start;
-      let newStart = Math.max(0, dragStartVal.current.start + deltaSec);
-      // Snap start and end
-      const excludes = [dragStartVal.current.start, dragStartVal.current.end];
-      const snappedStart = snapToNearest(newStart, excludes);
-      const snappedEnd = snapToNearest(newStart + dur, excludes);
-      if (snappedStart.snapped && (!snappedEnd.snapped || Math.abs(snappedStart.time - newStart) <= Math.abs(snappedEnd.time - (newStart + dur)))) {
-        newStart = snappedStart.time;
-        setSnapLineX(snappedStart.time * pxPerSec);
-      } else if (snappedEnd.snapped) {
-        newStart = snappedEnd.time - dur;
-        setSnapLineX(snappedEnd.time * pxPerSec);
-      } else {
-        setSnapLineX(null);
-      }
-      updateGlobalSub(dragIndex, { start: newStart, end: newStart + dur });
-    } else if (dragMode === 'sub-trim-left') {
-      let newStart = Math.max(0, dragStartVal.current.start + deltaSec);
-      const excludes = [dragStartVal.current.start];
-      const snap = snapToNearest(newStart, excludes);
-      if (snap.snapped) {
-        newStart = snap.time;
-        setSnapLineX(snap.time * pxPerSec);
-      } else {
-        setSnapLineX(null);
-      }
-      if (newStart < dragStartVal.current.end - MIN_CLIP_DURATION) {
-        updateGlobalSub(dragIndex, { start: newStart });
-      }
-    } else if (dragMode === 'sub-trim-right') {
-      let newEnd = Math.max(dragStartVal.current.start + MIN_CLIP_DURATION, dragStartVal.current.end + deltaSec);
-      const excludes = [dragStartVal.current.end];
-      const snap = snapToNearest(newEnd, excludes);
-      if (snap.snapped) {
-        newEnd = snap.time;
-        setSnapLineX(snap.time * pxPerSec);
-      } else {
-        setSnapLineX(null);
-      }
-      updateGlobalSub(dragIndex, { end: newEnd });
-    } else if (dragMode === 'bgm-move') {
-      const bgm = bgmClips[dragIndex];
-      if (!bgm) return;
-      let newStart = Math.max(0, dragStartVal.current.start + deltaSec);
-      const dur = bgm.duration;
-      const excludes = [dragStartVal.current.start, dragStartVal.current.start + dur];
-      const snappedStart = snapToNearest(newStart, excludes);
-      const snappedEnd = snapToNearest(newStart + dur, excludes);
-      if (snappedStart.snapped && (!snappedEnd.snapped || Math.abs(snappedStart.time - newStart) <= Math.abs(snappedEnd.time - (newStart + dur)))) {
-        newStart = snappedStart.time;
-        setSnapLineX(snappedStart.time * pxPerSec);
-      } else if (snappedEnd.snapped) {
-        newStart = snappedEnd.time - dur;
-        setSnapLineX(snappedEnd.time * pxPerSec);
-      } else {
-        setSnapLineX(null);
-      }
-      newStart = Math.max(0, newStart);
-      updateBgmClip(dragIndex, { start: newStart });
-    } else if (dragMode === 'bgm-trim-left') {
-      const bgm = bgmClips[dragIndex];
-      if (!bgm) return;
-      let newStart = Math.max(0, dragStartVal.current.start + deltaSec);
-      const snap = snapToNearest(newStart, [dragStartVal.current.start]);
-      if (snap.snapped) {
-        newStart = snap.time;
-        setSnapLineX(snap.time * pxPerSec);
-      } else {
-        setSnapLineX(null);
-      }
-      const delta = newStart - dragStartVal.current.start;
-      const origAudioStart = dragStartVal.current.audioStart ?? 0;
-      let newAudioStart = Math.max(0, origAudioStart + delta);
-      const newDuration = Math.max(MIN_CLIP_DURATION, dragStartVal.current.end - delta);
-      // 오리지널 길이를 초과하지 않도록 제한
-      const origDur = bgm.totalDuration || bgmWaveformDurations.current.get(bgm.source) || 0;
-      const maxDuration = origDur > 0 ? origDur - newAudioStart : Infinity;
-      updateBgmClip(dragIndex, { start: newStart, audioStart: newAudioStart, duration: Math.min(newDuration, maxDuration) });
-    } else if (dragMode === 'bgm-trim-right') {
-      let newDuration = Math.max(MIN_CLIP_DURATION, dragStartVal.current.end + deltaSec);
-      const bgm = bgmClips[dragIndex];
-      if (bgm) {
-        // 오리지널 길이를 초과하지 않도록 제한
-        const origDur = bgm.totalDuration || bgmWaveformDurations.current.get(bgm.source) || 0;
-        const maxDuration = origDur > 0 ? origDur - (bgm.audioStart || 0) : Infinity;
-        newDuration = Math.min(newDuration, maxDuration);
-        const newEnd = bgm.start + newDuration;
-        const snap = snapToNearest(newEnd, [bgm.start + dragStartVal.current.end]);
-        if (snap.snapped) {
-          newDuration = Math.max(MIN_CLIP_DURATION, snap.time - bgm.start);
-          newDuration = Math.min(newDuration, maxDuration);
-          setSnapLineX(snap.time * pxPerSec);
-        } else {
-          setSnapLineX(null);
-        }
-      }
-      updateBgmClip(dragIndex, { duration: newDuration });
-    }
-  }, [dragMode, dragIndex, pxPerSec, clipMeta, clipStarts, clipDurations, clips, globalSubs, bgmClips, getTimeFromX, seekToTime, updateClip, moveClip, moveClips, moveClipToTrackAndTime, selectedClipIndices, updateGlobalSub, updateBgmClip, snapToNearest]);
-
-  const handleMouseUp = useCallback(() => {
-    setDragMode(null);
-    setDragIndex(-1);
-    setSnapLineX(null);
-  }, []);
-
-  useEffect(() => {
-    if (dragMode) {
-      window.addEventListener('mousemove', handleMouseMove);
-      window.addEventListener('mouseup', handleMouseUp);
-      return () => {
-        window.removeEventListener('mousemove', handleMouseMove);
-        window.removeEventListener('mouseup', handleMouseUp);
-      };
-    }
-  }, [dragMode, handleMouseMove, handleMouseUp]);
-
-  // Keyboard: Delete, Undo/Redo, Copy/Paste, Fullscreen
-  useEffect(() => {
-    const onKey = (e: KeyboardEvent) => {
-      const active = document.activeElement;
-      const isInput = active && (active.tagName === 'INPUT' || active.tagName === 'TEXTAREA' || active.tagName === 'SELECT' || (active as HTMLElement).isContentEditable);
-
-      if (e.key === 'Delete' || e.key === 'Backspace') {
-        if (isInput) return;
-        const state = useEditorStore.getState();
-        if (state.activePanel === 'clip' && state.selectedClipIndices.length > 1) {
-          e.preventDefault();
-          state.removeClips(state.selectedClipIndices);
-        } else if (selectedClipIndex >= 0 && state.activePanel === 'clip') {
-          e.preventDefault();
-          removeClip(selectedClipIndex);
-        } else if (selectedSubIndex >= 0 && state.activePanel === 'subtitle') {
-          e.preventDefault();
-          removeSubtitle(selectedSubIndex);
-        } else if (selectedBgmIndex >= 0 && state.activePanel === 'bgm') {
-          e.preventDefault();
-          state.removeBgmClip(selectedBgmIndex);
-          setSelectedBgmIndex(-1);
-        }
-      }
-
-      // Undo: Cmd+Z
-      if (e.key === 'z' && (e.metaKey || e.ctrlKey) && !e.shiftKey) {
-        if (isInput) return;
-        e.preventDefault();
-        undo();
-      }
-
-      // Redo: Cmd+Shift+Z
-      if (e.key === 'z' && (e.metaKey || e.ctrlKey) && e.shiftKey) {
-        if (isInput) return;
-        e.preventDefault();
-        redo();
-      }
-
-      // Copy: Cmd+C
-      if (e.key === 'c' && (e.metaKey || e.ctrlKey)) {
-        if (isInput) return;
-        const state = useEditorStore.getState();
-        if (state.activePanel === 'subtitle' && state.selectedSubIndex >= 0) {
-          e.preventDefault();
-          copySubtitle();
-        } else if (state.activePanel === 'clip' && state.selectedClipIndex >= 0) {
-          e.preventDefault();
-          copyClip();
-        }
-      }
-
-      // Paste: Cmd+V
-      if (e.key === 'v' && (e.metaKey || e.ctrlKey)) {
-        if (isInput) return;
-        e.preventDefault();
-        paste();
-      }
-
-      // Fullscreen: F
-      if (e.key === 'f' || e.key === 'F') {
-        if (isInput) return;
-        // Dispatch custom event for fullscreen toggle
-        window.dispatchEvent(new CustomEvent('studio-fullscreen-toggle'));
-      }
-    };
-    window.addEventListener('keydown', onKey);
-    return () => window.removeEventListener('keydown', onKey);
-  }, [selectedClipIndex, selectedSubIndex, removeClip, removeSubtitle, undo, redo, copySubtitle, copyClip, paste]);
-
-  // Feature 4: Playhead auto-follow during playback
+  // Playhead auto-follow during playback
   useEffect(() => {
     if (!isPlaying || !scrollContainerRef.current) return;
     const container = scrollContainerRef.current;
@@ -487,184 +143,6 @@ export const Timeline: React.FC = () => {
       setPxPerSec(pxPerSec + delta);
     }
   }, [pxPerSec, setPxPerSec]);
-
-  const startDrag = (mode: DragMode, index: number, e: React.MouseEvent, startVal: { start: number; end: number; audioStart?: number }) => {
-    e.stopPropagation();
-    e.preventDefault();
-    setDragMode(mode);
-    setDragIndex(index);
-    dragStartX.current = e.clientX;
-    dragStartVal.current = startVal;
-  };
-
-  // BGM waveform cache
-  const bgmWaveforms = useRef<Map<string, Float32Array>>(new Map());
-  const bgmWaveformDurations = useRef<Map<string, number>>(new Map());
-  const clipWaveforms = useRef<Map<string, Float32Array>>(new Map());
-  const [clipAudioInfo, setClipAudioInfo] = useState<Record<string, ClipAudioInfo>>({});
-  const [waveformVersion, setWaveformVersion] = useState(0);
-
-  const buildMediaUrl = useCallback((source: string) => {
-    // 오디오 파일은 프록시 디렉토리에 없으므로 직접 경로 사용
-    const isAudio = /\.(mp3|wav|ogg|m4a|aac|flac)$/i.test(source);
-    if (isAudio) {
-      return `${getEditorConfig().apiUrl}/${encodeURIComponent(source)}`;
-    }
-    const prefix = getEditorConfig().mediaProxyPrefix || '/_proxy';
-    return `${getEditorConfig().apiUrl}${prefix}/${encodeURIComponent(source)}`;
-  }, []);
-
-  const extractWaveform = useCallback(async (source: string) => {
-    const url = buildMediaUrl(source);
-    console.log('[Waveform] Fetching:', url);
-    try {
-      const response = await fetch(url);
-      if (!response.ok) { console.warn('[Waveform] Fetch failed:', response.status); return null; }
-      const buf = await response.arrayBuffer();
-      console.log('[Waveform] ArrayBuffer size:', buf.byteLength);
-      const audioCtx = new AudioContext();
-      try {
-        const decoded = await audioCtx.decodeAudioData(buf);
-        const data = decoded.getChannelData(0);
-        console.log('[Waveform] Decoded samples:', data.length, 'sampleRate:', decoded.sampleRate);
-        // Very high resolution: store min/max pairs per window for detailed waveform
-        const targetLen = Math.min(16000, data.length);
-        const windowSize = Math.max(1, Math.floor(data.length / targetLen));
-        const result = new Float32Array(targetLen);
-        for (let i = 0; i < targetLen; i++) {
-          const start = i * windowSize;
-          const end = Math.min(start + windowSize, data.length);
-          let peak = 0;
-          for (let j = start; j < end; j++) {
-            const abs = Math.abs(data[j]);
-            if (abs > peak) peak = abs;
-          }
-          result[i] = peak;
-        }
-        const audioDuration = decoded.duration;
-        console.log('[Waveform] Generated', result.length, 'peaks, duration:', audioDuration);
-        return { peaks: result, duration: audioDuration };
-      } finally {
-        void audioCtx.close();
-      }
-    } catch (err) {
-      console.error('[Waveform] Error:', err);
-      return null;
-    }
-  }, [buildMediaUrl]);
-
-  const drawWaveformBars = useCallback((ctx: CanvasRenderingContext2D, widthPx: number, h: number, waveData: Float32Array | undefined, color: string, fallbackColor: string) => {
-    if (!waveData || waveData.length === 0) {
-      ctx.strokeStyle = fallbackColor;
-      ctx.lineWidth = 1;
-      ctx.beginPath();
-      for (let x = 0; x < widthPx; x += 3) {
-        const y = h / 2 + Math.sin(x * 0.15) * (h * 0.22);
-        x === 0 ? ctx.moveTo(x, y) : ctx.lineTo(x, y);
-      }
-      ctx.stroke();
-      return;
-    }
-
-    // Draw detailed mirrored waveform with gradient intensity
-    const mid = h / 2;
-    // Parse base color for gradient
-    const baseAlpha = 0.7;
-    for (let x = 0; x < widthPx; x++) {
-      const sampleIdx = Math.floor(x * waveData.length / widthPx);
-      const sampleEnd = Math.min(Math.floor((x + 1) * waveData.length / widthPx), waveData.length);
-      let peak = 0;
-      let rms = 0;
-      let count = 0;
-      for (let j = sampleIdx; j < sampleEnd; j++) {
-        const v = Math.abs(waveData[j]);
-        if (v > peak) peak = v;
-        rms += v * v;
-        count++;
-      }
-      rms = count > 0 ? Math.sqrt(rms / count) : 0;
-      // Outer bar (peak) — lighter
-      const peakH = Math.max(1, peak * (h - 2));
-      ctx.fillStyle = color;
-      ctx.fillRect(x, mid - peakH / 2, 1, peakH);
-      // Inner bar (RMS) — brighter, shows average energy
-      const rmsH = Math.max(1, rms * (h - 2));
-      ctx.fillStyle = color.replace(/[\d.]+\)$/, `${baseAlpha})`);
-      ctx.fillRect(x, mid - rmsH / 2, 1, rmsH);
-    }
-  }, []);
-
-  const drawBgmWaveform = useCallback((canvas: HTMLCanvasElement | null, bgmSource: string, widthPx: number, audioStart?: number, duration?: number, totalDuration?: number) => {
-    if (!canvas) return;
-    const ctx = canvas.getContext('2d');
-    if (!ctx) return;
-    const h = canvas.height;
-    ctx.clearRect(0, 0, widthPx, h);
-    let waveData = bgmWaveforms.current.get(bgmSource);
-    // Slice waveform to show only the audible portion (audioStart ~ audioStart+duration)
-    const audioDur = totalDuration || bgmWaveformDurations.current.get(bgmSource) || 0;
-    if (waveData && waveData.length > 0 && audioDur > 0 && audioStart != null && duration != null && duration < audioDur - 0.01) {
-      const startRatio = audioStart / audioDur;
-      const endRatio = Math.min((audioStart + duration) / audioDur, 1);
-      const startIdx = Math.floor(startRatio * waveData.length);
-      const endIdx = Math.min(Math.ceil(endRatio * waveData.length), waveData.length);
-      if (endIdx > startIdx) {
-        waveData = waveData.slice(startIdx, endIdx);
-      }
-    }
-    drawWaveformBars(ctx, widthPx, h, waveData, '#10b98166', '#10b98144');
-  }, [drawWaveformBars]);
-
-  const drawClipWaveform = useCallback((canvas: HTMLCanvasElement | null, clipSource: string, widthPx: number, muted: boolean) => {
-    if (!canvas) return;
-    const ctx = canvas.getContext('2d');
-    if (!ctx) return;
-    const h = canvas.height;
-    ctx.clearRect(0, 0, widthPx, h);
-    drawWaveformBars(
-      ctx,
-      widthPx,
-      h,
-      clipWaveforms.current.get(clipSource),
-      muted ? 'rgba(239,68,68,0.28)' : 'rgba(96,165,250,0.45)',
-      muted ? 'rgba(239,68,68,0.22)' : 'rgba(96,165,250,0.22)',
-    );
-  }, [drawWaveformBars]);
-
-  // Try loading waveform data for BGM clips
-  useEffect(() => {
-    for (const bgm of bgmClips) {
-      if (bgmWaveforms.current.has(bgm.source)) continue;
-      extractWaveform(bgm.source)
-        .then((result) => {
-          if (result) {
-            bgmWaveforms.current.set(bgm.source, result.peaks);
-            bgmWaveformDurations.current.set(bgm.source, result.duration);
-            setWaveformVersion((v) => v + 1);
-          }
-        })
-        .catch(() => { /* ignore */ });
-    }
-  }, [bgmClips, extractWaveform]);
-
-  useEffect(() => {
-    for (const clip of clips) {
-      if (clipAudioInfo[clip.source]) continue;
-      fetch(`${getEditorConfig().apiUrl}/api/media/probe/${encodeURIComponent(clip.source)}`)
-        .then((r) => r.ok ? r.json() : null)
-        .then((info) => {
-          if (!info) return;
-          setClipAudioInfo((prev) => prev[clip.source] ? prev : { ...prev, [clip.source]: { hasAudio: info.hasAudio !== false } });
-          if (info.hasAudio === false || clipWaveforms.current.has(clip.source)) return;
-          return extractWaveform(clip.source).then((result) => {
-            if (result) clipWaveforms.current.set(clip.source, result.peaks);
-          });
-        })
-        .catch(() => {
-          setClipAudioInfo((prev) => prev[clip.source] ? prev : { ...prev, [clip.source]: { hasAudio: true } });
-        });
-    }
-  }, [clips, clipAudioInfo, extractWaveform]);
 
   // Context menu for clip split
   const [ctxMenu, setCtxMenu] = useState<{ x: number; y: number; clipIdx: number; splitTime: number } | null>(null);
@@ -689,6 +167,7 @@ export const Timeline: React.FC = () => {
     zIndex: 2,
   });
 
+  // Subtitle lane assignment
   const subtitleLanes: number[] = [];
   const subtitleLaneEnds: number[] = [];
   for (let i = 0; i < globalSubs.length; i++) {
@@ -740,7 +219,6 @@ export const Timeline: React.FC = () => {
       )}
       {/* Toolbar */}
       <div style={{ display: 'flex', alignItems: 'center', gap: 6, padding: '4px 8px', borderBottom: '1px solid #222', flexShrink: 0 }}>
-        {/* Undo/Redo buttons */}
         <button
           className="be-btn"
           style={{ fontSize: 9, padding: '2px 6px', opacity: canUndo() ? 1 : 0.3 }}
@@ -875,9 +353,8 @@ export const Timeline: React.FC = () => {
       <div ref={scrollContainerRef} style={{ flex: 1, display: 'flex', overflow: 'auto' }}>
         {/* Track labels */}
         <div style={{ width: HEADER_W, flexShrink: 0, borderRight: '1px solid #222' }}>
-          {/* Video tracks: 트랙3(상단) → 트랙2 → 트랙1(하단) */}
           {Array.from({ length: NUM_VIDEO_TRACKS }, (_, ti) => {
-            const trackNum = NUM_VIDEO_TRACKS - ti; // 3, 2, 1
+            const trackNum = NUM_VIDEO_TRACKS - ti;
             return (
               <div key={`vtrack-${trackNum}`} style={{ height: TRACK_HEIGHT, display: 'flex', alignItems: 'center', paddingLeft: 8, fontSize: 9, color: '#888', borderBottom: '1px solid #1a1a1a' }}>
                 트랙 {trackNum}
@@ -901,9 +378,8 @@ export const Timeline: React.FC = () => {
           style={{ flex: 1, position: 'relative', minWidth: timelineWidth }}
           onClick={handleTimelineClick}
         >
-          {/* Clip tracks (multi-track: 트랙2=상단, 트랙0=하단) */}
+          {/* Clip tracks (multi-track) */}
           <div style={{ height: TRACK_HEIGHT * NUM_VIDEO_TRACKS, position: 'relative' }}>
-            {/* 트랙 구분선 */}
             {Array.from({ length: NUM_VIDEO_TRACKS }, (_, ti) => (
               <div key={`track-line-${ti}`} style={{ position: 'absolute', top: ti * TRACK_HEIGHT + TRACK_HEIGHT - 1, left: 0, right: 0, height: 1, background: '#1a1a1a' }} />
             ))}
@@ -948,7 +424,6 @@ export const Timeline: React.FC = () => {
                     if (!selectedClipIndices.includes(i)) {
                       setSelectedClipIndex(i);
                     }
-                    // dragOffsetSec: 마우스와 클립 왼쪽 간 오프셋
                     const mouseTime = getTimeFromX(e.clientX);
                     dragOffsetSec.current = mouseTime - (clip.timelineStart ?? clipStarts[i]);
                     startDrag('clip-move', i, e, { start: clip.start, end: clip.end });
@@ -1093,7 +568,7 @@ export const Timeline: React.FC = () => {
             })}
           </div>
 
-          {/* Subtitle track — auto lane assignment for overlapping subs */}
+          {/* Subtitle track */}
           <div style={{ height: subtitleTrackHeight, position: 'relative', borderBottom: '1px solid #1a1a1a', opacity: subsEnabled ? 1 : 0.3 }}>
             {globalSubs.map((sub, i) => {
               const left = sub.start * pxPerSec;
@@ -1259,7 +734,7 @@ export const Timeline: React.FC = () => {
             </div>
           )}
 
-          {/* Snap indicator line (Feature 3) */}
+          {/* Snap indicator line */}
           {snapLineX !== null && (
             <div
               style={{
